requirements.txt
django==4.1.4
psycopg2-binary==2.9.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
python-dotenv==0.21.0
pillow==9.3.0
djangorestframework_simplejwt==4.8.0
djoser==2.1.0
django-filter==22.1
drf-spectacular==0.25.1
pytest-factoryboy == 2.5.1
pytest-django==4.5.2
